Address feedback (refactor shims to avoid duplication)
